add lib.rs and move e2e_tests to tests directory
