+++ tests/e2e_tests.rs
@@ -1,21 +1,20 @@
-use super::http_api::HttpApi;
-use super::http_workers::HttpWorker;
-use super::identity;
-use super::identity::{Identity, Signer};
-use super::processor;
-use super::proxy::ProxyWorker;
-use super::redis;
-use super::storage::{ProxyStorage, RedisStorage, Storage};
-use super::twin::{Twin, TwinDB};
-use super::types::Message;
-use crate::http_api::UploadConfig;
-use anyhow::{Context, Result};
+use anyhow::{bail, Context, Result};
 use bb8_redis::redis::AsyncCommands;
+use rmb_rs::http_api::HttpApi;
+use rmb_rs::http_api::UploadConfig;
+use rmb_rs::http_workers::HttpWorker;
+use rmb_rs::identity;
+use rmb_rs::identity::{Identity, Signer};
+use rmb_rs::processor;
+use rmb_rs::proxy::ProxyWorker;
+use rmb_rs::redis;
+use rmb_rs::storage::{ProxyStorage, RedisStorage, Storage};
+use rmb_rs::twin::{Twin, TwinDB};
+use rmb_rs::types::Message;
 use sp_core::crypto::Pair;
 use sp_core::{ed25519::Pair as EdPair, sr25519::Pair as SrPair};
 use std::collections::HashMap;
 use std::panic;
-use std::path::PathBuf;
 use std::process::Command;
 use std::sync::Once;
 use tokio::time::{sleep, Duration};
