--- tests/mod.rs
@@ -1,4 +1,0 @@
-#[test]
-fn sample_test() {
-    assert_eq!(1 + 1, 2);
-}
